requirements.txt
sqlalchemy 
psycopg2-binary 
pandas
fastapi 
uvicorn